+++ package.json
@@ -8,7 +8,11 @@
   "license": "GPL-3.0-or-later",
   "private": false,
   "scripts": {
-    "lint": "yarn eslint --fix \"./src/**.ts\" -c .eslintrc.json"
+    "lint": "yarn eslint --fix \"./src/**.ts\" -c .eslintrc.json",
+    "clean": "rm -r build",
+    "build": "tsc",
+    "start": "node build/index.js",
+    "serve": "ts-node-dev src/index.ts"
   },
   "devDependencies": {
     "@types/bwip-js": "^2.0.0",
@@ -21,6 +25,7 @@
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1",
+    "ts-node-dev": "^1.0.0",
     "typescript": "^4.0.5"
   },
   "dependencies": {
